提交图集

+++ Assets/Script/Event/SO/IntEventSO.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 [CreateAssetMenu(menuName = "Event/IntEvent")]
-public class IntEventSO : BaseEventSO<int> 
+public class IntEventSO : BaseEventSO<float> 
 {
 	
 }
