+++ Assets/Script/Var/IntVariable.cs
@@ -4,12 +4,12 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Var/Int")]
 public class IntVariable : ScriptableObject 
 {
-	public int maxValue;
-	public int currentValue;
+	public float maxValue;
+	public float currentValue;
 
 	[CanBeNull]public IntEventSO ValueChangedEvent;
 
-	public void SetValue(int value)
+	public void SetValue(float value)
 	{
 		currentValue = value;
 		ValueChangedEvent?.RaiseEvent(value,this);
